put toggleModal logic in custom hook
