+++ src/pages/Landing.jsx
@@ -3,27 +3,27 @@ import { DinoDataContext } from "../context/DinoDataContext";
 import Hero from "../components/Hero";
 import Location from "../components/Location/Location";
 import LinksDrawer from "../components/Location/LinksDrawer/LinksDrawer";
-import { useState } from "react";
 import Chart from "./Chart";
 import styles from "./Landing.module.css";
-import { getCountries } from "../helpers/getCountries";
+import { getCountries } from "../components/Location/getCountries";
+import useToggleModal from "../components/Location/LinksDrawer/useToggleModal";
 
 export default function Landing() {
-  const [isLinksModalOpen, setIsLinksModalOpen] = useState(false);
-  const [country, setCountry] = useState();
-
   const { dinoData, loading, error } = useContext(DinoDataContext);
+
+  const {
+    isLinksModalOpen,
+    setIsLinksModalOpen,
+    country,
+    setCountry,
+    toggleModal,
+  } = useToggleModal();
 
   console.log(dinoData);
 
   if (loading) return <p>loading...</p>;
 
   if (error) return <p>cannot fetch dinosaurs</p>;
-
-  const toggleModal = (dino) => {
-    setCountry(dino.position.country);
-    setIsLinksModalOpen(true);
-  };
 
   return (
     <div>
